requirements.txt
flask==3.0.0
werkzeug==3.0.1
opencv-python-headless==4.8.1.78
numpy==1.26.2
pytesseract==0.3.10
pillow==10.1.0
